refactor!: change code to a more reusable structure for be able to work with others frameworks
