--- app/assets/config/oas_rails_manifest.js
@@ -1,1 +1,0 @@
-//= link_directory ../stylesheets/oas_rails .css
