--- app/helpers/oas_rails/application_helper.rb
@@ -1,4 +1,0 @@
-module OasRails
-  module ApplicationHelper
-  end
-end
