--- app/helpers/oas_rails/oas_rails_helper.rb
@@ -1,4 +1,0 @@
-module OasRails
-  module OasRailsHelper
-  end
-end
