--- app/helpers/oas_rails/test_helper.rb
@@ -1,4 +1,0 @@
-module OasRails
-  module TestHelper
-  end
-end
