--- app/jobs/oas_rails/application_job.rb
@@ -1,4 +1,0 @@
-module OasRails
-  class ApplicationJob < ActiveJob::Base
-  end
-end
