--- app/mailers/oas_rails/application_mailer.rb
@@ -1,6 +1,0 @@
-module OasRails
-  class ApplicationMailer < ActionMailer::Base
-    default from: "from@example.com"
-    layout "mailer"
-  end
-end
