--- app/views/layouts/oas_rails/application.html.erb
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= OasRails.config.info.title %></title>
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-    <meta charset="utf-8">
-    <!-- Important: rapi-doc uses utf8 characters -->
-  </head>
-  <body>
-    <%= yield %>
-  </body>
-</html>
