--- app/views/oas_rails/test/show.html.erb
@@ -1,1 +1,0 @@
-<%= @routes %>
