--- config/routes.rb
@@ -1,3 +1,0 @@
-OasRails::Engine.routes.draw do
-  get '(.:format)', to: 'oas_rails#index'
-end
