+++ lib/oas_rails/rails/controllers/oas_rails/application_controller.rb
