+++ lib/oas_rails/rails/controllers/oas_rails/oas_rails_controller.rb
