+++ lib/oas_rails/rails/views/oas_rails/oas_rails/index.html.erb
