+++ lib/oas_rails/spec/specification.rb
@@ -4,13 +4,9 @@ module OasRails
   module Spec
     class Specification
       include Specable
-      attr_accessor :components, :info, :openapi, :servers, :tags, :external_docs, :paths
+      attr_accessor :components, :info, :openapi, :servers, :tags, :external_docs, :paths, :route_extractor
 
-      # Initializes a new Specification object.
-      # Clears the cache if running in the development environment.
       def initialize
-        clear_cache unless Rails.env.production?
-
         @components = Components.new(self)
         @info = OasRails.config.info
         @openapi = '3.1.0'
@@ -18,20 +14,6 @@ module OasRails
         @tags = OasRails.config.tags
         @external_docs = {}
         @paths = Spec::Paths.new(self)
-      end
-
-      def build(route_extractor: Extractors::RouteExtractor)
-        route_extractor.host_paths.each do |path|
-          @paths.add_path(path)
-        end
-      end
-
-      # Clears the cache for MethodSource and RouteExtractor.
-      #
-      # @return [void]
-      def clear_cache
-        MethodSource.clear_cache
-        Extractors::RouteExtractor.clear_cache
       end
 
       def oas_fields
